有符号无符号数比较告警修复

+++ window_scene/screen_session_manager/src/screen_setting_helper.cpp
@@ -26,9 +26,9 @@ sptr<SettingObserver> ScreenSettingHelper::dpiObserver_;
 sptr<SettingObserver> ScreenSettingHelper::castObserver_;
 sptr<SettingObserver> ScreenSettingHelper::rotationObserver_;
 constexpr int32_t PARAM_NUM_TEN = 10;
-constexpr int32_t EXPECT_SCREEN_MODE_SIZE = 2;
-constexpr int32_t EXPECT_RELATIVE_POSITION_SIZE = 3;
-constexpr int32_t EXPECT_RESOLUTION_SIZE = 3;
+constexpr uint32_t EXPECT_SCREEN_MODE_SIZE = 2;
+constexpr uint32_t EXPECT_RELATIVE_POSITION_SIZE = 3;
+constexpr uint32_t EXPECT_RESOLUTION_SIZE = 3;
 constexpr uint32_t DATA_SIZE_INVALID = 0xffffffff;
 const std::string SCREEN_SHAPE = system::GetParameter("const.window.screen_shape", "0:0");
 
@@ -239,7 +239,7 @@ bool ScreenSettingHelper::SplitString(std::vector<std::string>& splitValues, con
 
 bool ScreenSettingHelper::IsNumber(const std::string& str)
 {
-    uint32_t length = str.size();
+    int32_t length = static_cast<int32_t>(str.size());
     int32_t start = 0;
     while (start < length && isspace(str[start])) {
         start++;
@@ -272,7 +272,7 @@ uint32_t ScreenSettingHelper::GetDataFromString(MultiScreenRecoverOption& option
     SplitString(splitValues, inputString, delimiter);
     std::string value;
     uint32_t dataSize = splitValues.size();
-    int32_t index = 0;
+    uint32_t index = 0;
     if (index < dataSize) {
         value = splitValues[index];
         if (!IsNumber(value)) {
